requirements.txt
instabot==0.70.0
python-dotenv==0.10.3
requests==2.22.0
